jacoco coverage setup.Robolectric is not providing coverage numbers, so added unit test that does
