+++ app/src/main/java/com/example/joshskeen/myapplication/MyActivity.java
@@ -6,20 +6,24 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
-public class MyActivity extends Activity {
+public class MyActivity extends Activity
+{
 
     Button mClickMeButton;
     TextView mHelloWorldTextView;
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
+    protected void onCreate(Bundle savedInstanceState)
+    {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_my);
         mHelloWorldTextView = (TextView) findViewById(R.id.helloWorldTextView);
         mClickMeButton = (Button) findViewById(R.id.clickMeBtn);
-        mClickMeButton.setOnClickListener(new View.OnClickListener() {
+        mClickMeButton.setOnClickListener(new View.OnClickListener()
+        {
             @Override
-            public void onClick(View v) {
+            public void onClick(View v)
+            {
                 mHelloWorldTextView.setText("HEY WORLD");
             }
         });
