+++ app/src/test/java/com/example/joshskeen/myapplication/RobolectricMyActivityUnitTest.java
@@ -1,0 +1,41 @@
+package com.example.joshskeen.myapplication;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.robolectric.Robolectric;
+import org.robolectric.RobolectricGradleTestRunner;
+import org.robolectric.annotation.Config;
+
+import static org.assertj.android.api.Assertions.assertThat;
+
+
+@RunWith(RobolectricGradleTestRunner.class)
+@Config(constants = BuildConfig.class, sdk = 21)
+public class RobolectricMyActivityUnitTest
+{
+
+    private MyActivity mActivity;
+
+    @Before
+    public void setup()
+    {
+        mActivity = Robolectric.buildActivity(MyActivity.class).create().get();
+    }
+
+    @Test
+    public void roboElectricMyActivityAppearsAsExpectedInitially()
+    {
+        assertThat(mActivity.mClickMeButton).hasText("Click me!");
+        assertThat(mActivity.mHelloWorldTextView).hasText("Hello world!");
+    }
+
+    @Test
+    public void roboElectricClickingClickMeButtonChangesHelloWorldText()
+    {
+        assertThat(mActivity.mHelloWorldTextView).hasText("Hello world!");
+        mActivity.mClickMeButton.performClick();
+        assertThat(mActivity.mHelloWorldTextView).hasText("HEY WORLD");
+    }
+
+}
